requirements.txt
hurry.filesize==0.9
hurry==1.0
PyYAML==5.1
